:white_check_mark: Make the API-convention gate see factory-registered kernels
`tests/api_conventions.py` enumerated kernel scope by walking `decorator_list`
(`_is_kernel`, `_kernel_scope_functions`). A kernel factory (CLAUDE.md section 2's
`wp.kernel(_k, name=...)`) registers a plain nested `def` whose body is Warp's DSL
but which carries no decorator, so every check keyed on that enumeration skipped it
entirely. This is section 4's own warning -- "a new Warp construct can silently
switch off a static check that predates it" -- with the construct being the factory,
and it went unnoticed because the factories postdate the checks.

Measured with an `ast` scan applying checks 16/17/18/20/22 to factory bodies: 14 such
bodies exist in the tree and 6 carried defects the family exists to catch.

  kernels/neighbors.py:568, :760   `tid = wp.tid()`                      check 22
  kernels/reduce.py:324, :702      `i = wp.tid()`                        check 22
  kernels/reduce.py:208, :209      `TILE_2D if ... else remaining`       check 20

Both ternaries have both arms already evaluated, so `wp.where` is a straight
substitution (section 1's caveat does not apply).

Check 13 was in the same blind spot and additionally walked only `tree.body`, so a
nested body was doubly invisible; extending it surfaced one more -- `kernels/reduce.py`'s
four factories name their output `out` where the convention is `out_`, now `out_result`.

The enumeration is fixed rather than the sites alone: `_factory_registered` collects the
names passed as the first argument to `wp.kernel(...)` / `wp.func(...)`, `_is_kernel`
takes an optional tree to consult it, and `_kernel_scope_functions` unions it in. The
enclosing factory stays out of scope, which is what keeps a `row_size: int` parameter
from tripping check 18.

The transferable half is the trigger: any enumeration keyed on `decorator_list` is a
candidate blind spot the moment a `wp.kernel(...)` / `wp.func(...)` *call* appears.

+++ triwarp/kernels/neighbors.py
@@ -565,7 +565,7 @@ def _bvh_nearest_row_kernel(row_size: int, name: str):
         out_indices: wp.array2d[wp.int32],
         out_distances: wp.array2d[wp.float32],
     ) -> None:
-        tid = wp.tid()
+        tid = wp.int32(wp.tid())
         q = queries[tid]
         row_indices = vec_indices()
         row_distances = vec_distances()
@@ -757,7 +757,7 @@ def _hashgrid_nearest_row_kernel(row_size: int, name: str):
         out_indices: wp.array2d[wp.int32],
         out_distances: wp.array2d[wp.float32],
     ) -> None:
-        tid = wp.tid()
+        tid = wp.int32(wp.tid())
         q = queries[tid]
         row_indices = vec_indices()
         row_distances = vec_distances()
